requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
coverage==7.9.1
    # via -r requirements/requirements.in
django==5.2.3
    # via
    #   -r requirements/requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.7.0
    # via -r requirements/requirements.in
django-environ==0.12.0
    # via -r requirements/requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r requirements/requirements.in
exceptiongroup==1.3.0
    # via pytest
flake8==7.2.0
    # via
    #   -r requirements/requirements.in
    #   flake8-html
flake8-html==0.4.3
    # via -r requirements/requirements.in
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via flake8-html
jsonschema==4.24.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.10
    # via -r requirements/requirements.in
pycodestyle==2.13.0
    # via flake8
pyflakes==3.3.2
    # via flake8
pygments==2.19.1
    # via
    #   flake8-html
    #   pytest
pytest==8.4.1
    # via -r requirements/requirements.in
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.3
    # via django
tomli==2.2.1
    # via pytest
typing-extensions==4.14.0
    # via
    #   asgiref
    #   exceptiongroup
    #   referencing
uritemplate==4.2.0
    # via drf-spectacular
